Form bug fix
added function out from component
added  custom schemas
added label
