+++ src/components/App.jsx
@@ -1,4 +1,5 @@
-import ModalProvider from 'hooks/ModalContext';
+import { useLayoutEffect } from 'react';
+
 import ClientSay from './ClientSay/ClientSay';
 import Count from './Count/Count';
 import Footer from './Footer/Footer';
@@ -7,14 +8,12 @@ import Hero from './Hero/Hero';
 import Preload from './Preload/Preload';
 import WhyMe from './WhyMe/WhyMe';
 import { Modal } from './Modal/Modal';
+import ModalProvider from 'hooks/ModalContext';
 import { instance } from 'helpers/axios';
-import { useLayoutEffect } from 'react';
 
 export const App = () => {
   useLayoutEffect(() => {
-    instance.get('/wakeup').catch(error => {
-      console.error();
-    });
+    instance.get('/wakeup');
   }, []);
 
   return (
